+++ packages/account-api-sdk/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sophon-labs/account-api-sdk
 
+## 2.1.1
+
+### Patch Changes
+
+- Support for refresh token on react sdk without react context
+- Updated dependencies
+  - @sophon-labs/account-core@2.1.1
+
 ## 2.1.0
 
 ### Minor Changes