+++ packages/account-react-native/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @sophon-labs/account-react-native
 
+## 2.1.1
+
+### Patch Changes
+
+- Support for refresh token on react sdk without react context
+- Updated dependencies
+  - @sophon-labs/account-message-bridge@2.1.1
+  - @sophon-labs/account-communicator@2.1.1
+  - @sophon-labs/account-provider@2.1.1
+  - @sophon-labs/account-core@2.1.1
+
 ## 2.1.0
 
 ### Minor Changes